[feat] Add translation settings and UI components
This commit adds translation settings into the application settings store, and introduces new UI components for `input`, `select`, `textarea`, and a translation form. It also arrives with the re-structural changes that adjust certain notification settings being moved into the `translation_settings` property. Note that a settings form for translation settings and a form control for translation strings are also included.